add example of how to extract command from user reply
When I started to use this adapter I found the basic examples a great start. But in the more advanced code with callback it only shows if (command === "...") but it does not explain or point in the direction of how to get "command" as string. This should help other beginners.
Best regards!