Add possibility to process in scripts the RAW data send by bot via botSendRaw.
It's useful, when in the same time you ned to process messageID, chatID and user.
Currently `botSendMessageId` and `botSendChatId` states updated in series, and to
process it in one callback it required some specific in coding.
In case of botSendRaw state you can process it more clear way.
I used the same config parameter, as for requestRaw enablement, to not overload
the adapter by config params.

Changes to be committed:
	modified:   io-package.json
	modified:   main.js