Don't send message on object change
This is a follow-up to eada2fcac9dbe452fd1ea03c90dacdc684846c69 (#190). Even with "Report changes only" checked, there used to be a message whenever the object had changed. That's because the `lastState` vanished and the next state update would be considered a change.

Fixed by reading the current value similar to how it's done in `readAllNames()`.

As a (positive?) side-effect, there are also no messages anymore when the monitoring is initially enabled or settings (like on/off text) are changed.